Figured out the actual difference between variables with initializations, assignment rules, and rate rules, and updated the code accordingly:  one formula is the initialization or asnt rule, and one formula is the rate rule, which can also have an initialization.  Worked out the logic of synchronizing variables of those three types, too. Added an enum, and updated the API accordingly.
As a result of this, all DNA elements are now either reaction rates or 
have assignment rules.  Nothing is constant (though it may effectively 
be if set to be a double.  May have to relax this later.)

Refactored some of the antimony-export and sbml-import code to eliminate 
duplication.

Updated some error messages.

Moved the 'examples' directory to be in the 'doc' directory; added the 
'biomodels' directory and the files in it.