No longer try to shorten the Antimony version of functions that use 'global' variables:  all variables for functions are treated as local variables when importing.  Writing a model with a function with a 'global' variable is still OK, but this form of the model will no longer round-trip.
The reason for the change is that Biomodels 79 failed to import any more, since a species was used as an argument to a function that needed to be turned into a modifier species reference.  Could have worked around it, but decided the shortening was too confusing in the first place, and we'd be better off not doing it.