Fix parsing of '-o' in formulas.
The interaction '-o' can be confused in formulas where you might have '-omega' (an actual case in Biomodels).

This doesn't fix the problem of having an actual 'o' parameter and wanting to subtract it in a formula:  I take '-o ' and '-o;' as meaning the interaction; hopefully this heuristic makes most things work.  (It at least fixes the '-omega' problem.)