Update docs
Fixes #120, and also includes initial conditions for all models that use the model/end format.  Also fixes a few typos, and syncs with some documentation apparently added to the Tellurium version of this documentation, but hadn't made it here.

Translated versions now created with pandoc; the .md version is officially canonical.